Removed a typo found by webcheck in backward-compatability.tex
Makefile.am: removed the commenting on document-hompage.tex on list of files,
corrected cable-theory to cable-theory-disgression but for some reason this 
gives an error. Will fix later.

Started to add a line for fixing the pdf links to the bog entries in contents-level1
but doesn't appear to work even tho the regex to test for it works. Working on this.